Move all common code into common_lib.py
